Utils: Refactor functions
Moved fit and loss to model, where it belongs semantically.
Improved function calls in notebook and function definitions.
Changed learning rate to match deepmind's implementation.
